--- a/pytest.ini
+++ b/pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 # Prevent helper project tests being collected as tests for pytest-rts
-norecursedirs = pytest_rts/tests/helper_project
+norecursedirs = pytest_rts/tests/helper_project pytest_rts/models