--- a/pytest_rts/models/base.py
+++ b/pytest_rts/models/base.py
@@ -1,0 +1,3 @@
+from sqlalchemy.ext.declarative import declarative_base
+
+Base = declarative_base()