--- a/pytest_rts/pytest/collect_plugin.py
+++ b/pytest_rts/pytest/collect_plugin.py
@@ -1,7 +1,6 @@
 """This module contains code for collecting newly added tests"""
 # pylint: disable=too-few-public-methods
-import sqlite3
-from pytest_rts.plugin import DB_FILE_NAME
+from pytest_rts.connection import MappingConn
 from pytest_rts.utils.testgetter import TestGetter
 
 
@@ -11,8 +10,7 @@ class CollectPlugin:
     def __init__(self):
         """Query existing test functions from database before collection"""
         self.collected = set()
-        self.conn = sqlite3.connect(DB_FILE_NAME)
-        self.testgetter = TestGetter(self.conn)
+        self.testgetter = TestGetter(MappingConn.session())
         self.existing_tests = self.testgetter.existing_tests
         self.testgetter.delete_newly_added_tests()
 
@@ -23,5 +21,5 @@ class CollectPlugin:
             if item.nodeid not in self.existing_tests:
                 self.collected.add(item.nodeid)
         self.testgetter.set_newly_added_tests(self.collected)
-        self.conn.commit()
-        self.conn.close()
+        MappingConn.session().commit()
+        MappingConn.session().close()