Hey @alexey-vyskubov 

I did a mistake in our profiling session and we were actually running the `master` version of pytest-rts when we were seeing similar results between using `SQLAlchemy` and the current `SQLite3` version. This version here seems to run 2x slower than `master` but the logic seems identical to `master`. I tried to improve performance with various tweaks to `session.flush` ordering and relationship management (orm objects vs id's) but I didn't really progress anywhere. Also, the changing database size between runs applies to the `master` version as well. I'll have to investigate why this happens.

With this version, the run of `pandas/tests/arithmetic` (as in CHANGELOG) takes nearly 12 minutes, instead of the 5.5 minutes. Also here's profiling between `master` and this version for "flask"

this branch:
```
         34149048 function calls (33673277 primitive calls) in 33.472 seconds

   Ordered by: internal time
   List reduced from 10699 to 25 due to restriction <25>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    19451    1.154    0.000    2.804    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/inspect.py:726(getmodule)
2202603/2202505    0.788    0.000    0.824    0.000 {built-in method builtins.hasattr}
4454118/4439711    0.600    0.000    0.651    0.000 {built-in method builtins.isinstance}
    27762    0.532    0.000    1.291    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/sqlalchemy/engine/default.py:771(_init_compiled)
   192382    0.464    0.000    0.464    0.000 {built-in method posix.lstat}
   108906    0.448    0.000    0.658    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:334(normpath)
   245828    0.424    0.000    0.653    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:71(join)
  2058213    0.403    0.000    0.549    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/inspect.py:63(ismodule)
    36758    0.381    0.000    0.381    0.000 {method 'execute' of 'sqlite3.Connection' objects}
   167350    0.378    0.000    0.505    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/sqlalchemy/orm/persistence.py:478(_collect_insert_commands)
     3262    0.307    0.000    0.307    0.000 {method 'executemany' of 'sqlite3.Cursor' objects}
     5582    0.300    0.000    0.300    0.000 {method 'executemany' of 'sqlite3.Connection' objects}
 19204/53    0.294    0.000   33.479    0.632 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/pluggy/callers.py:157(_multicall)
    24521    0.270    0.000    0.270    0.000 {method 'execute' of 'sqlite3.Cursor' objects}
  2234130    0.270    0.000    0.270    0.000 {method 'get' of 'dict' objects}
    35723    0.263    0.000    0.576    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/sqlalchemy/sql/elements.py:965(__init__)
    75556    0.263    0.000    0.263    0.000 {built-in method posix.stat}
   156531    0.260    0.000    0.260    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/sqlalchemy/sql/compiler.py:709(construct_params)
    27780    0.256    0.000    2.957    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/sqlalchemy/engine/base.py:1186(_execute_context)
28149/28146    0.255    0.000    1.385    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:396(_joinrealpath)
    22719    0.253    0.000    3.343    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/sqlalchemy/orm/persistence.py:1039(_emit_insert_statements)
        9    0.227    0.025    0.227    0.025 {built-in method posix.waitpid}
      540    0.225    0.000    0.225    0.000 {method 'executescript' of 'sqlite3.Connection' objects}
    11899    0.211    0.000    4.576    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/sqlalchemy/orm/session.py:2559(_flush)
     5768    0.207    0.000    0.207    0.000 {built-in method builtins.compile}

```

master:
```
         19609466 function calls (19378647 primitive calls) in 14.235 seconds

   Ordered by: internal time
   List reduced from 8071 to 25 due to restriction <25>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    19451    0.768    0.000    1.896    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/inspect.py:726(getmodule)
    16401    0.548    0.000    0.548    0.000 {method 'executemany' of 'sqlite3.Connection' objects}
1588066/1588058    0.521    0.000    0.550    0.000 {built-in method builtins.hasattr}
   189814    0.410    0.000    0.410    0.000 {built-in method posix.lstat}
   108435    0.387    0.000    0.568    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:334(normpath)
    60570    0.382    0.000    0.382    0.000 {method 'execute' of 'sqlite3.Connection' objects}
   242664    0.353    0.000    0.551    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:71(join)
3268012/3253587    0.306    0.000    0.316    0.000 {built-in method builtins.isinstance}
  1543336    0.277    0.000    0.375    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/inspect.py:63(ismodule)
 19204/53    0.264    0.000   14.114    0.266 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/pluggy/callers.py:157(_multicall)
    73758    0.229    0.000    0.229    0.000 {built-in method posix.stat}
     5780    0.227    0.000    0.227    0.000 {built-in method builtins.compile}
        9    0.223    0.025    0.223    0.025 {built-in method posix.waitpid}
27914/27911    0.218    0.000    1.202    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:396(_joinrealpath)
      540    0.205    0.000    0.205    0.000 {method 'executescript' of 'sqlite3.Connection' objects}
      410    0.204    0.000    0.204    0.000 /home/eero/Desktop/flask/src/flask/app.py:1980(try_trigger_before_first_request_functions)
126016/122040    0.161    0.000    3.749    0.000 {built-in method builtins.next}
    10819    0.159    0.000    0.340    0.000 /home/eero/Desktop/pytest-rts/.venv/lib/python3.8/site-packages/coverage/sqldata.py:879(arcs)
  1508602    0.140    0.000    0.141    0.000 {method 'get' of 'dict' objects}
   639957    0.125    0.000    0.125    0.000 {method 'startswith' of 'str' objects}
   126203    0.117    0.000    0.237    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/ast.py:214(iter_child_nodes)
      540    0.110    0.000    1.684    0.003 /home/eero/Desktop/pytest-rts/pytest_rts/utils/mappinghelper.py:238(save_testrun_data)
    18371    0.108    0.000    1.988    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/inspect.py:772(findsource)
    23267    0.107    0.000    0.646    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:449(relpath)
   189814    0.107    0.000    0.538    0.000 /home/eero/.pyenv/versions/3.8.5/lib/python3.8/posixpath.py:164(islink)

```